// apps/downloader/src/modules/facebook/types/facebook-video-metadata-dto.options.ts
export type FacebookVideoMetadataDTOOptions = {
  readonly title: string;
  readonly description: string;
  readonly uploader: string;
  readonly thumbnail: string;
  readonly viewCount: number;
  readonly duration: number;
  readonly id: string;
  readonly webpageUrl: string;
  readonly durationString: string;
};